feat(cliente): adiciona tratamento de exceção para caso tentar deletar cliente que já possui uma locação, estava dando erro nos testes e tive que implementar para mostrar ao usuario o porque dos erros.

+++ src/main/java/tech/ada/projeto_ada/poo1/cliente/service/DeletarClienteService.java
@@ -1,20 +1,28 @@
 package tech.ada.projeto_ada.poo1.cliente.service;
 
 import org.springframework.stereotype.Service;
+import tech.ada.projeto_ada.poo1.cliente.exception.ClienteComLocacoesException;
 import tech.ada.projeto_ada.poo1.cliente.repository.ClienteRepository;
+import tech.ada.projeto_ada.poo1.locacao.repository.LocacaoRepository;
 
 @Service
 public class DeletarClienteService {
     private final ClienteRepository repository;
     private final BuscarClienteService buscarClienteService;
+    private final LocacaoRepository locacaoRepository;
 
-    public DeletarClienteService(ClienteRepository repository, BuscarClienteService buscarClienteService) {
+
+    public DeletarClienteService(ClienteRepository repository, BuscarClienteService buscarClienteService, LocacaoRepository locacaoRepository) {
         this.repository = repository;
         this.buscarClienteService = buscarClienteService;
+        this.locacaoRepository = locacaoRepository;
     }
 
     public void deletarCliente(Long id) {
         buscarClienteService.buscarClientePorId(id);
+        if (locacaoRepository.existsByClienteId(id)) {
+            throw new ClienteComLocacoesException(id);
+        }
         repository.deleteById(id);
     }
